refact : 포트원 KG 이니시스 결제 v1 -> v2로 리팩터링

+++ src/main/java/com/example/bidly/domain/auction/dto/request/CreateAuctionRequest.java
@@ -10,7 +10,7 @@ import lombok.Getter;
 public class CreateAuctionRequest {
 
     private Long productId;
-    private Integer startPrice;
+    private Long startPrice;
     private AuctionDuration duration;
 
     public static CreateAuctionRequest from(ProductCreatedEvent event) {
